staff_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
requests==2.32.3
PyMySQL==1.1.1
